+++ src/main/java/racingcar/view/OutputView.java
@@ -1,7 +1,5 @@
 package racingcar.view;
 
-import racingcar.Winners;
-
 import java.util.List;
 
 import static racingcar.view.OutputConstant.RESULT_POSTFIX;
